breed_recognition/requirements.txt.txt
tk
pillow
opencv-python
numpy
tensorflow